+++ src/client.ts
@@ -1,2 +1,2 @@
 export type * from './types.js';
-export { ChannelType } from './consts.js';
+export * from './consts.js';